% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceDistributionPlot.r
\name{abundanceDistributionPlot}
\alias{abundanceDistributionPlot}
\title{abundanceDistributionPlot}
\usage{
abundanceDistributionPlot(
  x,
  m = NULL,
  fp = project.figuredirectory("bio.lobster"),
  fname
)
}
\arguments{
\item{\code{x}}{:the output file from nefsc.analysis or dfo.rv.analysis}

\item{\code{m}}{:the output of a loglinear model of abundance and distribution.}

\item{\code{fp}}{: file path for saving}

\item{\code{fn}}{: file name, needs to include .pdf as this is currently only setup to make pdf files.}
}
\value{
makes two plots of design weighted area occupied against N.yst and the time series of anomalies from the supplied log linear plot
}
\description{
uses the output of nefsc.analysis or dfo.rv.analysis to examine the relationship between design weighted area occupied and abundance
}
\examples{
require(bio.lobster)
aout= nefsc.analysis(DS='stratified.estimates.redo',p=p)
 aout$ln.n.Yst = log(aout$n.Yst)
 aout$ln.dwao = log(aout$dwao)
 model = lm(ln.dwao~ln.n.Yst,data=aout)
	abundanceDistributionPlot(aout,m=model'nefscSpringLFA41.pdf')
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
