% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnamR.r
\name{bnamR}
\alias{bnamR}
\title{bnamR
Read in and reshape the bnam model output file. This currentely requires a specific format of bnam output (from Brickman Oct 2019).}
\usage{
bnamR(
  redo = F,
  bnam_location = "~/Downloads/BNAM_Tbtm_1990_2018.mat",
  outfile = file.path(project.datadirectory("bio.lobster"), "data", "BNAM",
    "bnams.reshape.rds"),
  standard = T
)
}
\arguments{
\item{redo}{= if redo = T redo the reshape, else load in the previously reshaped file}

\item{bnam_location}{= location of the file (Brickman Oct 2019 structure)}

\item{outfile}{= where you want the output to live}
}
\description{
bnamR
Read in and reshape the bnam model output file. This currentely requires a specific format of bnam output (from Brickman Oct 2019).
}
\examples{
anc = bnam(redo = T)
}
\author{
Adam Cook
}
