% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaSwept.r
\name{calcAreaSwept}
\alias{calcAreaSwept}
\title{calcTrackDist}
\usage{
calcAreaSwept(
  sensors = "Marport",
  GearSpreadRange = list(NEST = c(5, 15), `280 BALLOON` = c(5, 25))
)
}
\description{
calcTrackDist
}
