% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.dd.dddd.r
\name{convert.dd.dddd}
\alias{convert.dd.dddd}
\title{convert.dd.dddd}
\usage{
convert.dd.dddd(x, format = "dec.deg")
}
\arguments{
\item{x}{vector of degrees-minutes-seconds or degree.minutes or degrees to convert to decimal degrees}

\item{format}{one of 'dec.deg' or 'deg.min' reflecting input data}
}
\description{
convert.dd.dddd
}
\author{
Brad Hubley
}

