% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPUEgam.R
\name{CPUEgam}
\alias{CPUEgam}
\title{CPUEgam}
\usage{
CPUEgam(data, mf)
}
\arguments{
\item{data}{is the data frame for fitting a generalized additive catch rate model. Data frame must contain columns of CPUE, SYEAR and DATE}

\item{mf}{is the gam model to be fitted}
}
\value{
returns a plot of CPUE data by date along with the fitted line as well as a data.frame with raw data and predicted
}
\description{
CPUEgam
}
\author{
Brad Hubley
}

