% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.setup.r
\name{db.setup}
\alias{db.setup}
\title{db.setup}
\usage{
db.setup(RODBC = F)
}
\value{
"con" which defines details of the Oracle connection. Structured differently for RODBC and ROracle.
}
\description{
This function allows for automatic selection of ROracle and RODBC packages as the default for Oracle database queries. ROacle is the default. If running 32-bit R or do not have ROracle installed, RODBC will be selected.
}
\author{
Ben Zisserson
}
