% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depricated_calcAreaSwept.r
\name{depricated_calcAreaSwept}
\alias{depricated_calcAreaSwept}
\title{calcTrackDist}
\usage{
depricated_calcAreaSwept(
  sensors = "Marport",
  GearSpreadRange = list(NEST = c(5, 15), `280 BALLOON` = c(5, 25))
)
}
\description{
calcTrackDist
}
