% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfo.rv.jonah.analysis.r
\name{dfo.rv.jonah.analysis}
\alias{dfo.rv.jonah.analysis}
\title{dfo.rv.jonah.analysis}
\usage{
dfo.rv.jonah.analysis(DS = "stratified.estimates",
  out.dir = "bio.lobster", p = p, ip = NULL, save = T)
}
\arguments{
\item{\code{DS}}{:the selection of analysis, options include \code{stratified.estimates}}

\item{\code{out.dir}}{: specify the location of data saves, default is null and uses the project.datadirectory function as default}

\item{\code{p}}{: the parameter list which contains the specifics of the analysis at a minimum includes the season and area for analysis}
}
\value{
saves or loads .rdata objects
}
\description{
Stratified analysis of DFO lobster data with bootstrapped resampling and set-up the data for sensitivity analysis
}
\examples{
require(devtools)
load_all('E:/git/LobsterScience/bio.lobster') #to load from directory rather than package if modifying
dfo.rv.analysis(DS = 'stratified.estimates')
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
