% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicusum.r
\name{dicusum}
\alias{dicusum}
\title{dicusum}
\usage{
dicusum(x, ref.index, k)
}
\arguments{
\item{\code{x}}{:vector of time series data}

\item{\code{ref.period}}{: index used for reference period}

\item{\code{k}}{: threshold differences to detect a change, also termed allowance}
}
\value{
returns a list of positive and negative deviations as well as reference index
}
\description{
decision interval CUSUM control chart analysis
}
\examples{
x = c(1,2,3,4,5,3,2,1,2,6)
dicusum(x,1:3,0.2)
}
\references{
Mesnil and Pettigas 2008 Aquat Living Resour 22 187-192
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
