% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findValue.r
\name{findValue}
\alias{findValue}
\title{findValue}
\usage{
findValue(x, y, ref = 0.4)
}
\arguments{
\item{x}{is the xaxis}

\item{y}{is the yaxis}

\item{ref}{is the level of y you want to find the corresponding x value for}
}
\value{
x value
}
\description{
Finds the corresponding x value to match a reference y value
}
