% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBerried.r
\name{getBerried}
\alias{getBerried}
\title{getBerried}
\usage{
getBerried(p)
}
\arguments{
\item{\code{p}}{:parameter list containing (at a minimum) lens (length bins), doy (day of year), temp (temperature at doy)}

\item{\code{gdd}}{: logical: growth using degree-days}
}
\value{
The probability of bearing eggs at length
}
\author{
Brad Hubley, \email{Brad.Hubley@dfo-mpo.gc.ca}
}
