% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroMat.r
\name{getGroMat}
\alias{getGroMat}
\title{getGroMat}
\usage{
getGroMat(p, gdd = F)
}
\arguments{
\item{\code{p}}{:parameter list containing (at a minimum) lens (length bins), doy (day of year), temp (temperature at doy)}

\item{\code{gdd}}{: logical: growth using degree-days}
}
\value{
The transition matrix for moulting and probability of moulting at length
}
\description{
Prediction of moulting based on day since last moult, size and can include temperature through a degree day metric
}
\author{
Brad Hubley, \email{Brad.Hubley@dfo-mpo.gc.ca}
}

