% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmCIs.r
\name{glmCIs}
\alias{glmCIs}
\title{glmCIs}
\usage{
glmCIs(model.output, newdata = NULL)
}
\arguments{
\item{model.output}{= the glm model output}

\item{newdata}{= the data on which to predict the glm model}
}
\value{
data frame containing the new data with predictions and CIs
}
\description{
This function estimates correct CIs for a glm model by estimating on the link and then backtransforming
}
\examples{
None yet
}
