% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundfish.db.r
\name{groundfish.db}
\alias{groundfish.db}
\title{groundfish.db.r}
\usage{
groundfish.db(DS = "gscat.odbc.redo", p = NULL, taxa = "all", datayrs = NULL)
}
\arguments{
\item{\code{DS}}{:the selection of analysis, options include \code{stratified.estimates}}

\item{\code{p}}{: the parameter list which contains the specifics of the analysis at a minimum includes the season and area for analysis}
}
\value{
saves or loads .rdata objects
}
\description{
data extractions from groundfish database
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
