% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILTS_ITQ_All_Data.r
\name{ILTS_ITQ_All_Data}
\alias{ILTS_ITQ_All_Data}
\title{ILTS_ITQ_All_Data}
\usage{
ILTS_ITQ_All_Data(
  x,
  species = 2550,
  redo_base_data = F,
  size = NULL,
  sex = NULL,
  aggregate = F,
  return_tow_tracks = F
)
}
\arguments{
\item{species}{Species codes from ISDB (2550 = lobster; 10 = Atlantic cod; etc)}

\item{redo_base_data}{if TRUE redoes all the swept area calculations and data filtering / cleanup}

\item{size}{specifies size range for your data; requires a two element vector c(min, max)}

\item{sex}{specifies sex for data; only applicable to lobster and crabs. 0 = unid; 1= male; 2= female; 3=egg carrying}

\item{aggregate}{if TRUE aggregates the data by trip and set based on above arguments. If no size or sex is specified its total abundance.}
}
\value{
Data objects that contain the data for use in further analyses.
}
\description{
This function pulls out all the ILTS ITQ data for all sets and species either aggregated by length and/or sex or raw data.
}
\examples{
ILTS_ITQ_All_Data(species=2550, size=c(1,200),sex=c(3),aggregate=T)
}
