% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la.r
\name{la}
\alias{la}
\title{la}
\usage{
la(package = "bio.lobster", base.loc = git.repo)
}
\arguments{
\item{\code{base.loc}}{:the base directory of bio.lobster}
}
\description{
runs the load_all script from devtools to load bio.lobster after function change
}
\examples{
la() 
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
