% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobsterLogsProcess.r
\name{LobsterLogsProcess}
\alias{LobsterLogsProcess}
\title{LobsterLogsProcess
Processing lobster logbook data to get to clean data in a useable form}
\usage{
LobsterLogsProcess(bumpup = F)
}
\arguments{
\item{bumpup}{is for using the slip data to bumpup logs as a percentage basis. This defaults to false.}
}
\value{
returns an R data object and saves the data to a .csv
}
\description{
LobsterLogsProcess
Processing lobster logbook data to get to clean data in a useable form
}
\note{
Make sure you run the lobster.db('logs.redo') to get the up to date data. Need to check fishing season dates as a appropriate filter.
}
\seealso{
lobster.db('logs.redo')
}
\author{
Brad Hubley & Adam Cook
}
