% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobsterMap.r
\name{LobsterMap}
\alias{LobsterMap}
\title{LobsterMap}
\usage{
LobsterMap(area = "custom", ylim = c(40, 52), xlim = c(-74, -47),
  save = F, output = "bio.lobster", fname = "one.pdf", mapRes = "HR",
  land.col = "wheat", title = "", nafo = NULL, boundaries = "LFAs",
  bathy.source = "topex", isobaths = seq(100, 1000, 100), bathcol = rgb(0,
  0, 1, 0.1), topolines = NULL, topocol = rgb(0.8, 0.5, 0, 0.2),
  points.lst = NULL, pt.cex = 1, lines.lst = NULL, poly.lst = NULL,
  contours = NULL, image.lst = NULL, color.fun = tim.colors, zlim,
  grid = NULL, stippling = F, lol = F, labels = "lfa", labcex = 1.5,
  LT = T, plot.rivers = T, addSummerStrata = F, addsubareas = F,
  subsetSummerStrata = NULL, addGeorgesStrata = F, addAmericanStrata = F,
  addGrids = T, ...)
}
\arguments{
\item{area}{= 'custom' where xlim & ylim are specified or select from area list below}

\item{mapRes}{= coastline detail ('LR' = low resolution, 'MR' = medium resolution, 'HR' = high resolution, 'UR' = ultra resolution)}

\item{title}{= plot title}

\item{boundaries}{= for ploting specific management boundaries}

\item{bathcol}{= isobath line color, default is transparent blue}

\item{topolines}{= plots topographic lines for specified elevations from topex data}

\item{points.lst}{= points to overlay on map in PBSmapping format - list with 2 elements: 1st element is eventSet (EID, POS, X, Y), 2nd element is eventData (EID, pch, col, etc.)}

\item{lines.lst}{= lines to overlay on map in PBSmapping format - list with 2 elements: 1st element is polySet (PID, SID, POS, X, Y), 2nd element is polyData (PID, SID, lty, col, etc.)}

\item{poly.lst}{= polygons to overlay on map in PBSmapping format - list with 2 elements: 1st element is polySet (PID, SID, POS, X, Y), 2nd element is polyData (PID, SID, border, col, etc.)}

\item{contours}{= plots overlaping polygons as contours (same format as poly.lst)}

\item{image.lst}{= image to overlay on map - list with 3 elements (x, y, z), 'bathymetry' produces image from bathymetry data}

\item{color.fun}{= color function for image}

\item{zlim}{= zlim for image}

\item{grid}{= size of grid in degrees, default is no grid}

\item{stippling}{= adds stippling to land (purely for visual effect)}

\item{lol}{= adds water colored border to coastline (purely for visual effect)}

\item{isobath}{= plots bathymetry lines for specified depths from topex data}

\item{bathcol}{= topolines line color, default is transparent brown}
}
\description{
LobsterMap
}
\examples{
LobsterMap(area='34')
catchgrids <-lobGridPlot(subset(logsInSeason,SYEAR==2007,c("LFA","GRID_NUM","WEIGHT_KG")),lvls=c(100,50000,100000,200000,400000,600000,800000,1000000),FUN=sum,border=NA)
LobsterMap('all',poly.lst=catchgrids[1:2],title="2007 Lobster Catch")
ContLegend("bottomright",lvls=catchgrids$lvls/1000,Cont.data=catchgrids,title="Catch (t)",inset=0.02,cex=0.8,bg='white')
}
\author{
Brad Hubley
}
