% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobsterMarport.r
\name{LobsterMarport}
\alias{LobsterMarport}
\title{LobsterMarport}
\usage{
LobsterMarport(file)
}
\arguments{
\item{file}{is the marport file obtained during the 2015 ILTS survey which contains the net sensor information as well as the gps data feeds}
}
\value{
returns a list containing two elements: 1) the trawl sensor information 2) the gps track of the file
}
\description{
LobsterMarport
}
\author{
Adam Cook
}
