% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobsterSurveyProcess.r
\name{LobsterSurveyProcess}
\alias{LobsterSurveyProcess}
\title{LobsterSurveyProcess}
\usage{
LobsterSurveyProcess(species = 2550, size.range = c(0, 200),
  lfa = "34", yrs, mths = c("May", "Jun", "Jul", "Aug", "Sep", "Oct"),
  gear.type = NULL, sex = c(1:3, NA), bin.size = 5, LFS = 160,
  Net = NULL, comparative = F, biomass = F)
}
\arguments{
\item{size.range}{defines the minimum and maximum value and is a filter (default is 0, 200mm CW)}

\item{lfa}{defines the specific LFA for the ILTS}

\item{yrs}{is the survey years to estimate}

\item{gear.type}{survey trawl net identification (if !NULL options are '280 BALLOON' or 'NEST')}

\item{bin.size}{aggregates the abundance into size bins (default is 5mm bins)}

\item{mnts}{months of the survey, defaults to the full year}
}
\value{
data.frame of survey data called 'surveyLobsters'
}
\description{
LobsterSurveyProcess
}
\author{
Brad Hubley & Manon Cassista-Da Ros
}
