% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePBS.r
\name{makePBS}
\alias{makePBS}
\title{makePBS}
\usage{
makePBS(x, polygon = T, projection = "LL", coords = NULL)
}
\description{
unknown
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\concept{abysmally documented}
