% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla.r
\name{mla}
\alias{mla}
\title{mla}
\usage{
mla(base.loc = "C:/Users/cassistadarosm/Documents/GitHub/")
}
\arguments{
\item{\code{base.loc}}{:the base directory of bio.lobster}
}
\description{
runs the load_all script from devtools for MCDR to load bio.lobster after function change
}
\examples{
la() 
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
