% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.db.r
\name{nefsc.db}
\alias{nefsc.db}
\title{nefsc.db}
\usage{
nefsc.db(DS = "odbc.dump.redo")
}
\arguments{
\item{\code{DS}}{= the selection of data, consists of a full data dump from ODBC account through \code{odbc.dump}. Or individual data tables can be rebuilt (with .redo) or loaded as 'uscat','usdet','usinf','ustrata.area'}
}
\value{
saves or loads .rdata object
}
\description{
Pulls the offshore lobster data from NEFSC trawl surveys
}
\examples{
nefsc.db(DS = 'odbc.dump.redo')
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}

