% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dicusum.r
\name{plot.dicusum}
\alias{plot.dicusum}
\title{plot.dicusum}
\usage{
plot.dicusum(x, x.index, h, ...)
}
\arguments{
\item{\code{x}}{:output of dicusum.r}

\item{\code{x.index}}{: index for x labels}

\item{\code{h}}{: decision interval}
}
\value{
returns a plot
}
\description{
figure decision interval CUSUM control chart analysis
}
\examples{
x = c(1,2,3,4,5,3,2,1,2,6)
a = dicusum(x,1:3,0.2)
plot.dicusum(a)
}
\references{
Mesnil and Pettigas 2008 Aquat Living Resour 22 187-192
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
