% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smooth_sdmtmb.R
\name{plot_smooth_sdmtmb}
\alias{plot_smooth_sdmtmb}
\title{Plot a smooth term from an sdmTMB model}
\usage{
plot_smooth_sdmtmb(
  object,
  select = 1,
  n = 100,
  level = 0.95,
  ggplot = FALSE,
  rug = TRUE,
  return_data = FALSE
)
}
\description{
Plot a smooth term from an sdmTMB model
}
