% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPrMolt.R
\name{pPrMolt}
\alias{pPrMolt}
\title{getIncr}
\usage{
pPrMolt(cw, a, b, d)
}
\arguments{
\item{\code{cw}}{: carapace width}

\item{\code{a}}{: molt probability parameter}

\item{\code{b}}{: molt probability parameter}

\item{\code{d}}{: days since last molt}
}
\value{
The predicted probability of moulting
}
\description{
Made up function that gives you probability of moulting based on day since last moult, size and can include temperature through a degree day metric. Needs to be calibrated with real data.
}
\author{
Brad Hubley, \email{Brad.Hubley@dfo-mpo.gc.ca}
}

