% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_with_metadata.r
\name{save_with_metadata}
\alias{save_with_metadata}
\title{Save an object with rich metadata (provenance + environment)}
\usage{
save_with_metadata(
  object,
  file,
  description = "",
  fn = NULL,
  dictionary = NULL,
  extra_meta = list(),
  compress = "gzip"
)
}
\arguments{
\item{object}{The R object to save (data.frame, list, model, etc.)}

\item{file}{Output .rds filename}

\item{description}{A short human-readable description}

\item{fn}{Optional: pass the function that produced `object` (for code capture)}

\item{dictionary}{Optional data dictionary}

\item{extra_meta}{Optional named list with any additional fields you want}

\item{compress}{One of "gzip" (default), "xz", "bzip2", or FALSE}
}
\description{
Save an object with rich metadata (provenance + environment)
}
\examples{
x <- data.frame(YEAR=1999:2013,Stat = runif(15,0,100))
dict = data.frame(
   Column = c("Year","Anomalies"),
   type  = sapply(x, function(x) paste(class(x), collapse = "/")))
save_with_metadata(object=x,file='SaveFile.rds',dictionary=dict)
}
