% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_chull.r
\name{sf_chull}
\alias{sf_chull}
\title{This function generates a convext hull around a set of points}
\usage{
sf_chull(x, buffer = 0.25)
}
\arguments{
\item{x}{is your sf point set to create a convex hull}

\item{buffer}{is your buffer dist (meaningless unless crs is UTM)}
}
\value{
a buffered polygon of all points
}
\description{
This function generates a convext hull around a set of points
}
