% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMolt.R
\name{simMolt}
\alias{simMolt}
\title{simMolt}
\usage{
simMolt(p, gdd = T, continuous.recruitment = F)
}
\arguments{
\item{\code{p}}{:parameter list containing (at a minimum) region, doy (day of year), temp (temperature at doy)}

\item{\code{gdd}}{: growing degree day}

\item{\code{continuous.recruitment}}{: T/F whether you want to add new recruits at startPop size at an annual time step}

\item{\code{recruitment.vector}}{: Variable Recruitment vector of length p$nt/round(365/p$timestep)}
}
\value{
The predicted probability of moulting
}
\description{
Prediction of moulting based on day since last moult, size and can include temperature through a degree day metric
}
\author{
Brad Hubley, \email{Brad.Hubley@dfo-mpo.gc.ca}
}
