% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLobsterMap.r
\name{ggLobsterMap}
\alias{ggLobsterMap}
\title{Title: Map LFAs
Description: Plot lobster fishing areas with labels, grids, bathymetry, etc.}
\usage{
ggLobsterMap(
  area = "custom",
  fill.colours = "grey",
  ylim = c(40, 52),
  xlim = c(-74, -47),
  LFAfill = "white",
  attrData = NULL,
  attrColumn = "Z",
  addGrids = T,
  addNAFO = F,
  nafo = "4X",
  bathy = T,
  fw = NULL,
  legLab = "",
  addLFAlines = T,
  addLFALabels = F,
  addGridLabels = F,
  addPoints = F,
  pts,
  addNAFOLabels = F,
  scaleTrans = "identity",
  brks = NULL,
  return.object = F,
  layerDir = file.path(bio.directory, "bio.lobster.data", "mapping_data"),
  LFA_label_size = 8,
  grid_label_size = 8,
  colourLFA = T,
  ...
)
}
\description{
Title: Map LFAs
Description: Plot lobster fishing areas with labels, grids, bathymetry, etc.
}
