% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLobsterMap.r
\name{ggLobsterMap}
\alias{ggLobsterMap}
\title{Title: Map LFAs
Description: Plot lobster fishing areas with labels, grids, bathymetry, etc.}
\usage{
ggLobsterMap(
  area = "custom",
  fill.colours = "grey",
  ylim = c(40, 52),
  xlim = c(-74, -47),
  LFAfill = "white",
  attrData = NULL,
  attrColumn = "Z",
  addGrids = FALSE,
  addNAFO = FALSE,
  nafo = "4X",
  bathy = TRUE,
  fw = NULL,
  legLab = "",
  addLFAlines = TRUE,
  addLFALabels = FALSE,
  addGridLabels = FALSE,
  addPoints = FALSE,
  pts = NULL,
  addNAFOLabels = FALSE,
  scaleTrans = "identity",
  brks = NULL,
  return.object = FALSE,
  layerDir = file.path(bio.directory, "bio.lobster.data", "mapping_data"),
  LFA_label_size = 4,
  grid_label_size = 3,
  show37label = FALSE,
  colourLFA = NULL,
  ...
)
}
\description{
Title: Map LFAs
Description: Plot lobster fishing areas with labels, grids, bathymetry, etc.
}
