% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobsterSurveyProcess.r
\name{LobsterSurveyProcess}
\alias{LobsterSurveyProcess}
\title{LobsterSurveyProcess}
\usage{
LobsterSurveyProcess(size.range = c(0, 220), lfa, yrs, mths = c("Jan",
  "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"),
  bin.size = 5)
}
\arguments{
\item{size.range}{defines the minimum and maximum value and is a filter (default is 0, 220mm CW).}

\item{lfa}{defines the specific LFA for the ILTS}

\item{yrs}{is the survey years to estimate}

\item{bin.size}{aggregates the abundance into size bins defaults to 5mm bins}

\item{mnts}{months of the survey, defaults to the full year}
}
\value{
data.frame of survey data
}
\description{
LobsterSurveyProcess
}
\author{
Brad Hubley
}

