% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.db.r
\name{nefsc.db}
\alias{nefsc.db}
\title{nefsc.db}
\usage{
nefsc.db(DS = "odbc.dump.redo", fn.root = NULL, p = p)
}
\arguments{
\item{\code{DS}}{:the selection of data, consists of a full data dump from ODBC account through \code{odbc.dump}. Or individual data tables can be rebuilt (with \code{.redo}) or loaded as \code{uscat},\code{usdet},\code{usinf},\code{usstrata.area}. Running any of the previous with the .clean arguement will clean up data, take care of vessel conversions and make it ready for further analysis}

\item{\code{fn.root}}{: specify the location of data saves, default is null and uses the project.datadirectory function as default}
}
\value{
saves or loads .rdata objects named \code{usinf}, \code{usdet}, \code{uscat}, \code{usstrat.area}
}
\description{
Pulls the offshore lobster data from NEFSC trawl surveys as well as the strata areas. Included in the strata areas are the proportion of the total strata area incorporated in each of the offshore lobster subunits for post stratification.
}
\examples{
require(devtools)
load_all('E:/git/LobsterScience/bio.lobster') #to load from directory rather than package if modifying
nefsc.db(DS = 'odbc.dump.redo')
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}

