% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPrMolt.R
\name{pPrMolt}
\alias{pPrMolt}
\title{pPrMolt}
\usage{
pPrMolt(p, cl)
}
\arguments{
\item{\code{cl}}{: carapace length}

\item{\code{p}}{: parameter list with moltPrModel and degreedays (ddoy)}
}
\value{
The predicted probability of moulting
}
\description{
probability of moulting based on degreedays since last moult, size.
}
\author{
Brad Hubley, \email{Brad.Hubley@dfo-mpo.gc.ca}
}
